refactored the functionalities + menu

+++ src/main/java/ro/pao/service/impl/GradeServiceImpl.java
@@ -5,6 +5,7 @@ import ro.pao.model.Grade;
 import ro.pao.repository.GradeRepository;
 import ro.pao.repository.impl.GradeRepositoryImpl;
 import ro.pao.service.GradeService;
+import ro.pao.service.StudentService;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -14,7 +15,7 @@ import java.util.logging.Level;
 
 public class GradeServiceImpl implements GradeService {
     private static final GradeRepository gradeRepository = new GradeRepositoryImpl();
-
+    private static final StudentService studentService = new StudentServiceImpl();
     @Override
     public Optional<Grade> getById(UUID id) {
         try {
@@ -37,6 +38,8 @@ public class GradeServiceImpl implements GradeService {
     public void addOnlyOne(Grade newObject) {
         try {
             gradeRepository.addNewObject(newObject);
+
+            studentService.updateStudentGradeList(newObject.getStudentId());
         } catch (SQLException e) {
             LogServiceImpl.getInstance().log(Level.SEVERE, e.getMessage());
         }
@@ -44,20 +47,45 @@ public class GradeServiceImpl implements GradeService {
 
     @Override
     public void addMany(List<Grade> objectList) {
-        try {
-            gradeRepository.addAllFromGivenList(objectList);
-        } catch (SQLException e) {
-            LogServiceImpl.getInstance().log(Level.SEVERE, e.getMessage());
+        for (Grade grade : objectList) {
+            addOnlyOne(grade);
         }
     }
 
     @Override
     public void removeById(UUID id) {
-        gradeRepository.deleteObjectById(id);
+        Optional<Grade> gradeOptional = getById(id);
+
+        if (gradeOptional.isPresent()) {
+            Grade grade = gradeOptional.get();
+
+            gradeRepository.deleteObjectById(id);
+
+            studentService.updateStudentGradeList(grade.getStudentId());
+        }
+    }
+
+    @Override
+    public List<Grade> getGradesByStudentId(UUID studentId) {
+        return gradeRepository.getAllGradesByStudentId(studentId);
     }
 
     @Override
     public void updateById(UUID id, Grade newObject) {
-        gradeRepository.updateObjectById(id, newObject);
+        Optional<Grade> gradeOptional = getById(id);
+
+        if (gradeOptional.isPresent()) {
+            Grade grade = gradeOptional.get();
+
+            gradeRepository.updateObjectById(id, newObject);
+
+            if (grade.getStudentId().toString().equals(newObject.getStudentId().toString())) {
+                studentService.updateStudentGradeList(grade.getStudentId());
+            } else {
+                studentService.updateStudentGradeList(grade.getStudentId());
+
+                studentService.updateStudentGradeList(newObject.getStudentId());
+            }
+        }
     }
 }
